requirements.txt
streamlit
snowflake-connector-python
pandas
python-dotenv
pytest
hypothesis[pandas]
